requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.3
dataclasses==0.8
decorator==4.4.2
dgl-cu110==0.6.1
idna==3.2
networkx==2.5.1
numpy==1.19.5
Pillow==8.3.1
requests==2.26.0
scipy==1.5.4
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
typing-extensions==3.10.0.0
urllib3==1.26.6
